Moving methods around to their logical places
